// api/reputation/model/comment-report-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Accommodation Service - Accommodation Owner APIs
 * Api Documentation for Accommodation Owner Users
 *
 * The version of the OpenAPI document: Accommodation Owner
 * Contact: amir.hossein.khalouei@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { MetaCommentReportResponseDto } from './meta-comment-report-response-dto';
import { OwnerCommentResponseDto } from './owner-comment-response-dto';

/**
 * 
 * @export
 * @interface CommentReportResponseDto
 */
export interface CommentReportResponseDto {
    /**
     * 
     * @type {MetaCommentReportResponseDto}
     * @memberof CommentReportResponseDto
     */
    meta?: MetaCommentReportResponseDto;
    /**
     * 
     * @type {Array<OwnerCommentResponseDto>}
     * @memberof CommentReportResponseDto
     */
    commentsTable?: Array<OwnerCommentResponseDto> | null;
}